+++ modules/client/src/addresslistVoting/internal/client/methods.ts
@@ -77,7 +77,7 @@ import {
   SubgraphAddresslistVotingProposal,
   SubgraphAddresslistVotingProposalListItem,
 } from "../types";
-import { InvalidActionsOrderError } from "@aragon/sdk-common";
+import { AlwaysFailingProposalError } from "@aragon/sdk-common";
 
 /**
  * Methods module the SDK Address List Client
@@ -96,8 +96,8 @@ export class AddresslistVotingClientMethods extends ClientCore
   ): AsyncGenerator<ProposalCreationStepValue> {
     const signer = this.web3.getConnectedSigner();
 
-    if (isFailingProposal(params.actions || [])) {
-      throw new InvalidActionsOrderError();
+    if (isFailingProposal(params.actions)) {
+      throw new AlwaysFailingProposalError();
     }
 
     const addresslistContract = AddresslistVoting__factory.connect(
