+++ modules/client/src/multisig/internal/client/methods.ts
@@ -72,7 +72,7 @@ import { toMultisigProposal, toMultisigProposalListItem } from "../utils";
 import { toUtf8Bytes } from "@ethersproject/strings";
 import { MultisigClientEncoding } from "./encoding";
 import { IMultisigClientMethods } from "../../interfaces";
-import { InvalidActionsOrderError } from "@aragon/sdk-common";
+import { AlwaysFailingProposalError } from "@aragon/sdk-common";
 
 /**
  * Methods module the SDK Address List Client
@@ -91,8 +91,8 @@ export class MultisigClientMethods extends ClientCore
   ): AsyncGenerator<ProposalCreationStepValue> {
     const signer = this.web3.getConnectedSigner();
 
-    if (isFailingProposal(params.actions || [])) {
-      throw new InvalidActionsOrderError();
+    if (isFailingProposal(params.actions)) {
+      throw new AlwaysFailingProposalError();
     }
 
     const multisigContract = Multisig__factory.connect(
